==> code/02_imputation.Rmd <==
---
title: "Imputation"
author: "Jae Yeon Kim"
output:
  html_document: 
    toc: true
    theme: united
---

# Setup 

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(
        remotes, # for installing packages on GitHub 
        mice, # for imputation
        miceadds, # additional functions for mice
        tidyverse, # for the tidyverse framework 
        naniar, # for missing values 
        ggthemes, # for fancy ggplot themes
        here, # for self-contained projects
        ggpubr # for pub-ready themes 
)

source(here("functions", "utils.r"))

remotes::install_github("njtierney/naniar")

# for publication-friendly theme 
theme_set(theme_pubr())
```

# Importing files 

Importing the file that we created in `01_data_cleaning.Rmd`.

```{r}
# Import 
cleaned <- read.csv(here("processed_data/cleaned.csv"))

cleaned$X <- NULL
```

# Examining missing responses

- Listwise deletion (`na.rm = TRUE`) works only if these observations are missing completely at random (MCAR). This is a strong assumption, as it happens very rarely. 

## Exploratory analysis (data visualization)

- The `nanair package` provides many valuable functions for inspecting missing data values. I used the `miss_var_summary() function` from this package to check the missing pattern and visualized it using ggplot2. 

```{r}
# Visualize the percentage of missing values by variables and subgroups 

cleaned %>%
    group_by(race) %>% # group by race
    miss_var_summary() %>% # summary 
    filter(n_miss != 0) %>% # show variables that include missing values
    ggplot(aes(x = reorder(variable, pct_miss), y = pct_miss)) +
        geom_col() +
        facet_wrap(~race) + # facet by race 
        coord_flip() + # flip axes 
        labs(x = "Variables",
             y = "Missing rate (%)")
ggsave(here("outputs/missing_rate.png"), height = 8)
```

# Multiple imputation 

In this case, I took a multiple imputation approach. I recovered (imputed) missing values based on the observed data. I did so by making simulations of these imputed values to garner some uncertainty measures.

## Imputing data 

In this project, I used the `mice package` developed by [Stef van Buuren](https://stefvanbuuren.name/).  The following code shows how the imputation model is set up. I comment on each argument to make it explicit how the model is set up. For instance, the `m argument` refers to the number of imputations, and I set it to `5`.

```{r}
imp <- mice(
  cleaned %>% select(-respid),
    seed = 1234, # for reproducibility
    m = 5, # the number of imputations
    maxit = 100, # the max number of iterations 
    method = "pmm", # predictive mean method
    print = FALSE) 
```

## Pooling results

```{r}
# Pooling results 
imputed <- mice::complete(imp) # the first imputed data

# Check 
paste("The number of missing values in the data after multiple imputation is", sum(is.na(imputed)))
```

# the imputed data

```{r}
imputed$respid <- cleaned$respid

write.csv(imputed, here("processed_data/imputed.csv"))

names(imputed)
```
